Added main files first developed with Deepnote
Two items. 
Item 1 - The "ALGO 1 - SOXL EMA_SMA Cross v_2" is the main algo decision engine that takes in live stock trading data. Every data point is the latest minute ticker price. In this case that is from $SOXL which is a triple leveraged semi-conductor etf. The criteria for a trade and target ticker(s) can be changed to suit an individual strategy. 

Item 2 - The package requirements for the code to run.